# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Runs API",
    "description": "Start a multi-step agent run, follow it live to completion (poll or SSE, resumable across disconnects), and read back what happened and what it cost. Every step is instrumented once into an append-only event log (the source of truth \u2014 this API never reads from OpenTelemetry) and into OTel spans/metrics for the paired trace and analytics story. See `trace_id` on the run envelope to jump into the corresponding waterfall trace.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Liveness check",
        "description": "Unversioned \u2014 used by container orchestration, not the v1 product surface.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Start a run",
        "description": "Durable-first: the run is persisted as `pending` *before* execution is spawned, so this `202` never lies \u2014 by the time you get a response, the run exists and is resumable even if the process crashes immediately after. Send an `Idempotency-Key` header to make retries of this call safe (PRD \u00a73.3): the same key + same body replays the original response with no new run created; the same key + a different body is a `409 idempotency_error`.",
        "operationId": "create_run_v1_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional. Makes this call safe to retry (PRD \u00a73.3).",
              "title": "Idempotency-Key"
            },
            "description": "Optional. Makes this call safe to retry (PRD \u00a73.3)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request (body/query/path params).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "`Idempotency-Key` was already used with a different request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List runs",
        "description": "Cursor-paginated (`{data, has_more, next_cursor}`); the ULID run ID doubles as the cursor, sorted `created_at desc`. Filters: `status`, `agent_id`, `metadata.<key>` (repeat the prefix per key, e.g. `?metadata.customer_id=cust_123`), `created_after`/`created_before`.",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a prior page's `next_cursor`.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a prior page's `next_cursor`."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "created_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created After"
            }
          },
          {
            "name": "created_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Before"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request (body/query/path params).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get a run",
        "description": "The run envelope, folded from the event log \u2014 status, running totals (tokens, cost_usd), and trace_id. Safe to poll at any point in the run's lifecycle.",
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Run not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request (body/query/path params).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/steps": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get a run's steps",
        "description": "Per-step state (type, status, attempt count, last_error, per-step tokens/cost). Always returned in full, never paginated \u2014 see `StepListResponse`.",
        "operationId": "get_steps_v1_runs__run_id__steps_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Run not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request (body/query/path params).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Cancel a run",
        "description": "Persists a `cancelling` transition and signals the in-process task. The runner stops at the next step boundary; a `finally` block guarantees a terminal `cancelled` event follows. Watch it happen live on the `/events` stream.",
        "operationId": "cancel_run_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Run not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The run is already in a terminal status (completed/failed/cancelled) \u2014 first terminal write wins.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request (body/query/path params).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Follow a run live (SSE)",
        "description": "Server-Sent Events tail of the persisted event log \u2014 never the runner directly, so this works identically whether the run is mid-execution or long finished. Historical events replay first, then new ones stream as they're appended; the connection closes after a terminal event. Reconnect with `Last-Event-ID: N` to resume from N+1, byte-identical to a fresh replay from that point. Idle connections get a `: heartbeat` comment every 15s to defeat proxy buffering timeouts.",
        "operationId": "stream_events_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "`text/event-stream` \u2014 one `id:`/`event:`/`data:` block per domain event (`run.started`, `step.started`, `step.completed`, \u2026), or `: heartbeat` on idle.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Run not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request (body/query/path params).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "invalid_request",
              "not_found",
              "conflict",
              "idempotency_error",
              "internal_error"
            ],
            "title": "Type",
            "description": "Closed set \u2014 stable enough for a client to branch on."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Growable specific case, e.g. `run_not_found`, `run_terminal`."
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param",
            "description": "The request field/header this error is attributed to, if any."
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Also attached to the trace \u2014 support-ticket-to-trace link"
          }
        },
        "type": "object",
        "required": [
          "type",
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "description": "The one error shape every non-2xx response uses, including validation\nerrors (FastAPI's default 422 handler is overridden to match).",
        "examples": [
          {
            "code": "unknown_agent_id",
            "message": "unknown agent_id: 'agent-mystery'. Valid: agent-researcher, agent-simple, agent-flaky",
            "param": "agent_id",
            "request_id": "01JZ9X8K3Q7YB2N4M6P8R0T2W4",
            "type": "invalid_request"
          }
        ]
      },
      "RunCreateRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "description": "Selects a behavior profile: agent-researcher, agent-simple, or agent-flaky.",
            "examples": [
              "agent-researcher"
            ]
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input",
            "description": "Arbitrary JSON payload, part of the deterministic recipe. \u226432KB JSON-encoded.",
            "examples": [
              {
                "prompt": "summarize the attached document"
              }
            ]
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed",
            "description": "Omit to have the server generate one (returned on the envelope) \u2014 every run, chosen or generated, is replayable from its recipe alone."
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Opaque string\u2192string tags. Filterable via `metadata.<key>` on the list endpoint and attached as span attributes."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "input"
        ],
        "title": "RunCreateRequest",
        "examples": [
          {
            "agent_id": "agent-researcher",
            "input": {
              "prompt": "summarize the attached document"
            },
            "metadata": {
              "customer_id": "cust_123"
            },
            "seed": 42
          }
        ]
      },
      "RunEnvelope": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "description": "Echoed back \u2014 server-generated if the request omitted it."
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "tokens_in": {
            "type": "integer",
            "title": "Tokens In"
          },
          "tokens_out": {
            "type": "integer",
            "title": "Tokens Out"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "description": "Rounded to 6dp; storage keeps full precision."
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Set at creation, before the runner starts \u2014 open this in Grafana/Tempo to see the run's span waterfall."
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunError"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "agent_id",
          "seed",
          "input",
          "metadata",
          "tokens_in",
          "tokens_out",
          "cost_usd",
          "trace_id",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "RunEnvelope",
        "description": "The run's current state, folded from the event log \u2014 never from the\nrunner or OTel (the store is the source of truth). Identical shape on\ncreation, on poll, and after cancel; only `status` and the totals change\nover the run's lifetime."
      },
      "RunError": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/RunErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "RunError"
      },
      "RunErrorCode": {
        "type": "string",
        "enum": [
          "step_failed",
          "interrupted_by_restart",
          "cancelled_by_user"
        ],
        "title": "RunErrorCode",
        "description": "Closed set of domain error codes the runner and service layer may attach."
      },
      "RunListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunEnvelope"
            },
            "type": "array",
            "title": "Data"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "data",
          "has_more",
          "next_cursor"
        ],
        "title": "RunListResponse"
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "cancelling",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus"
      },
      "StepListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/StepOut"
            },
            "type": "array",
            "title": "Data"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StepListResponse",
        "description": "Steps are always returned in full \u2014 a run's step count is small and\nbounded by its profile (PRD \u00a74: at most 8 top-level steps, one level of\nsub-agent nesting), so there's no cursor to advance. `has_more` is\nalways `false` and `next_cursor` always `null`; the shape is still the\nshared `{data, has_more, next_cursor}` list envelope (PRD \u00a73.3) rather\nthan a bare array, so a client never needs a special case to tell\nwhether a given list-shaped endpoint is paginated."
      },
      "StepOut": {
        "properties": {
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "parent_step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Step Id",
            "description": "Set only for a sub-agent's nested children (one level deep)."
          },
          "step_type": {
            "$ref": "#/components/schemas/StepType"
          },
          "status": {
            "$ref": "#/components/schemas/StepStatus"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt",
            "description": "1-indexed. >1 means this step retried at least once."
          },
          "tokens_in": {
            "type": "integer",
            "title": "Tokens In"
          },
          "tokens_out": {
            "type": "integer",
            "title": "Tokens Out"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "description": "Rounded to 6dp; storage keeps full precision."
          },
          "last_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunError"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "step_id",
          "parent_step_id",
          "step_type",
          "status",
          "attempt",
          "tokens_in",
          "tokens_out",
          "cost_usd",
          "last_error",
          "created_at",
          "updated_at"
        ],
        "title": "StepOut"
      },
      "StepStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ],
        "title": "StepStatus"
      },
      "StepType": {
        "type": "string",
        "enum": [
          "model_call",
          "tool_call",
          "sub_agent"
        ],
        "title": "StepType"
      }
    }
  },
  "tags": [
    {
      "name": "runs",
      "description": "Create, follow, cancel, and read back agent runs \u2014 the whole product surface."
    }
  ]
}